refactor(quick-math): remove legacy files and rename Framework to canonical names
