+++ src/lib/games/quick-math/game.ts
@@ -1,18 +1,59 @@
+import { BaseGame } from '@/lib/games/core/BaseGame'
 import type {
-    GameState,
-    GameConfig,
-    GameCallbacks,
-    MathQuestion,
-    GameStats,
-} from './types'
+    BaseGameState,
+    BaseGameConfig,
+    BaseGameCallbacks,
+    BaseGameStats,
+} from '@/lib/games/core/types'
+import type { GameID } from '@/lib/games'
 
-export class QuickMathGame {
-    private state: GameState
-    private config: GameConfig
-    private callbacks: GameCallbacks
-    private gameTimer: number | null = null
+// Quick Math specific types
+export interface MathQuestion {
+    id: string
+    question: string
+    answer: number
+    operation: 'addition' | 'subtraction'
+    operand1: number
+    operand2: number
+}
+
+export interface QuickMathState extends BaseGameState {
+    currentQuestion: MathQuestion | null
+    questionsAnswered: number
+    correctAnswers: number
+    incorrectAnswers: number
+    currentAnswer: string
+}
+
+export interface QuickMathConfig extends BaseGameConfig {
+    pointsPerCorrectAnswer: number
+    maxNumber: number
+    operations: Array<'addition' | 'subtraction'>
+}
+
+export interface QuickMathStats extends BaseGameStats {
+    totalQuestions: number
+    correctAnswers: number
+    incorrectAnswers: number
+    accuracy: number
+    averageTimePerQuestion: number
+}
+
+export interface QuickMathCallbacks extends BaseGameCallbacks {
+    onQuestionUpdate?: (question: MathQuestion) => void
+    onAnswerSubmit?: (
+        correct: boolean,
+        question: MathQuestion,
+        answer: string
+    ) => void
+}
+
+export class QuickMathFrameworkGame extends BaseGame<
+    QuickMathState,
+    QuickMathConfig,
+    QuickMathStats
+> {
     private questionStartTime: number = 0
-    // Achievement tracking flags
     private achievementFlags: {
         seenOnePlusOne: boolean
         onePlusOneIncorrect: boolean
@@ -25,25 +66,55 @@ export class QuickMathGame {
         zeroAnswerIncorrect: false,
     }
 
-    constructor(config: GameConfig, callbacks: GameCallbacks) {
-        this.config = config
-        this.callbacks = callbacks
-        this.state = this.getInitialState()
+    constructor(
+        gameId: GameID,
+        config: QuickMathConfig,
+        callbacks: QuickMathCallbacks = {}
+    ) {
+        super(gameId, config, callbacks, {
+            basePoints: config.pointsPerCorrectAnswer,
+        })
     }
 
-    private getInitialState(): GameState {
+    createInitialState(): QuickMathState {
         return {
             score: 0,
+            timeRemaining: this.config.duration,
+            isActive: false,
+            isPaused: false,
+            isGameOver: false,
+            gameStarted: false,
             currentQuestion: null,
             questionsAnswered: 0,
             correctAnswers: 0,
             incorrectAnswers: 0,
-            timeRemaining: this.config.gameDuration,
-            isGameActive: false,
-            isGameOver: false,
-            gameStartTime: null,
             currentAnswer: '',
         }
+    }
+
+    protected onGameStart(): void {
+        this.achievementFlags = {
+            seenOnePlusOne: false,
+            onePlusOneIncorrect: false,
+            seenOperand999: false,
+            zeroAnswerIncorrect: false,
+        }
+        this.generateNextQuestion()
+    }
+
+    protected onGameEnd(
+        _finalScore: number,
+        _finalStats: QuickMathStats
+    ): void {
+        // Game ended, cleanup if needed
+    }
+
+    protected onGameReset(): void {
+        this.state.currentQuestion = null
+        this.state.questionsAnswered = 0
+        this.state.correctAnswers = 0
+        this.state.incorrectAnswers = 0
+        this.state.currentAnswer = ''
     }
 
     private generateRandomQuestion(): MathQuestion {
@@ -61,7 +132,7 @@ export class QuickMathGame {
             question = `${operand1} + ${operand2}`
             answer = operand1 + operand2
         } else {
-            // For subtraction, ensure result is positive
+            // Ensure subtraction result is positive
             const larger = Math.max(operand1, operand2)
             const smaller = Math.min(operand1, operand2)
             question = `${larger} - ${smaller}`
@@ -78,164 +149,128 @@ export class QuickMathGame {
         }
     }
 
-    private startGameTimer(): void {
-        this.gameTimer = window.setInterval(() => {
-            this.state.timeRemaining--
-            this.callbacks.onTimeUpdate(this.state.timeRemaining)
-
-            if (this.state.timeRemaining <= 0) {
-                this.endGame()
-            }
-        }, 1000)
-    }
-
-    private stopGameTimer(): void {
-        if (this.gameTimer) {
-            clearInterval(this.gameTimer)
-            this.gameTimer = null
-        }
-    }
-
-    public startGame(): void {
-        this.state = this.getInitialState()
-        this.state.isGameActive = true
-        this.state.gameStartTime = Date.now()
+    private generateNextQuestion(): void {
+        const question = this.generateRandomQuestion()
+        this.state.currentQuestion = question
         this.questionStartTime = Date.now()
 
-        // Reset achievement flags
-        this.achievementFlags = {
-            seenOnePlusOne: false,
-            onePlusOneIncorrect: false,
-            seenOperand999: false,
-            zeroAnswerIncorrect: false,
+        if (
+            question.operation === 'addition' &&
+            question.operand1 === 1 &&
+            question.operand2 === 1
+        ) {
+            this.achievementFlags.seenOnePlusOne = true
+        }
+        if (question.operand1 === 999 || question.operand2 === 999) {
+            this.achievementFlags.seenOperand999 = true
         }
 
-        this.generateNextQuestion()
-        this.startGameTimer()
-        this.callbacks.onGameStart()
+        // Notify callbacks only
+        const callbacks = this.callbacks as QuickMathCallbacks
+        if (callbacks.onQuestionUpdate) {
+            callbacks.onQuestionUpdate(question)
+        }
     }
 
     public submitAnswer(answer: string): boolean {
-        if (!this.state.isGameActive || !this.state.currentQuestion) {
+        if (!this.state.isActive || !this.state.currentQuestion) {
             return false
         }
 
-        const numericAnswer = parseInt(answer, 10)
-        const isCorrect = numericAnswer === this.state.currentQuestion.answer
-
+        const isCorrect = parseInt(answer) === this.state.currentQuestion.answer
         this.state.questionsAnswered++
-
-        // Update achievement flags based on current question and answer
-        const q = this.state.currentQuestion
-        if (
-            q.operation === 'addition' &&
-            q.operand1 === 1 &&
-            q.operand2 === 1 &&
-            !isCorrect
-        ) {
-            this.achievementFlags.onePlusOneIncorrect = true
-        }
-        if (q.answer === 0 && !isCorrect) {
-            this.achievementFlags.zeroAnswerIncorrect = true
-        }
 
         if (isCorrect) {
             this.state.correctAnswers++
-            this.state.score += this.config.pointsPerCorrectAnswer
-            this.callbacks.onScoreUpdate(this.state.score)
+            this.addScore(this.config.pointsPerCorrectAnswer, 'correct_answer')
         } else {
             this.state.incorrectAnswers++
+            const q = this.state.currentQuestion
+            if (q) {
+                if (
+                    q.operation === 'addition' &&
+                    q.operand1 === 1 &&
+                    q.operand2 === 1
+                ) {
+                    this.achievementFlags.onePlusOneIncorrect = true
+                }
+                if (q.answer === 0) {
+                    this.achievementFlags.zeroAnswerIncorrect = true
+                }
+            }
         }
 
-        this.generateNextQuestion()
+        // Notify callbacks only
+        const callbacks = this.callbacks as QuickMathCallbacks
+        if (callbacks.onAnswerSubmit) {
+            callbacks.onAnswerSubmit(
+                isCorrect,
+                this.state.currentQuestion,
+                answer
+            )
+        }
+
+        // Generate next question if game is still active
+        if (this.state.isActive) {
+            this.generateNextQuestion()
+        }
+
         return isCorrect
     }
 
-    private generateNextQuestion(): void {
-        this.state.currentQuestion = this.generateRandomQuestion()
-        this.state.currentAnswer = ''
-        this.questionStartTime = Date.now()
-
-        // Update achievement flags when question appears
-        const q = this.state.currentQuestion
-        if (q) {
-            if (
-                q.operation === 'addition' &&
-                q.operand1 === 1 &&
-                q.operand2 === 1
-            ) {
-                this.achievementFlags.seenOnePlusOne = true
-            }
-            if (q.operand1 === 999 || q.operand2 === 999) {
-                this.achievementFlags.seenOperand999 = true
-            }
-        }
-        this.callbacks.onQuestionUpdate(this.state.currentQuestion)
+    public getCurrentQuestion(): MathQuestion | null {
+        return this.state.currentQuestion
     }
 
-    public updateCurrentAnswer(answer: string): void {
-        this.state.currentAnswer = answer
-    }
-
-    public getCurrentAnswer(): string {
-        return this.state.currentAnswer
-    }
-
-    public endGame(): void {
-        this.state.isGameActive = false
-        this.state.isGameOver = true
-        this.stopGameTimer()
-
-        const stats = this.getGameStats()
-        this.callbacks.onGameOver(this.state.score, stats)
-    }
-
-    public getGameStats(): GameStats {
-        const totalTime = this.state.gameStartTime
-            ? (Date.now() - this.state.gameStartTime) / 1000
-            : 0
+    public getGameStats(): QuickMathStats {
+        const totalTime = this.getTimerStatus().elapsedTime || 0
 
         return {
+            finalScore: this.state.score,
+            timeElapsed: Math.floor(totalTime),
+            gameCompleted: this.state.isGameOver,
             totalQuestions: this.state.questionsAnswered,
             correctAnswers: this.state.correctAnswers,
             incorrectAnswers: this.state.incorrectAnswers,
             accuracy:
                 this.state.questionsAnswered > 0
-                    ? (this.state.correctAnswers /
-                          this.state.questionsAnswered) *
-                      100
+                    ? Math.round(
+                          (this.state.correctAnswers /
+                              this.state.questionsAnswered) *
+                              100
+                      )
                     : 0,
             averageTimePerQuestion:
                 this.state.questionsAnswered > 0
-                    ? totalTime / this.state.questionsAnswered
+                    ? Math.round(
+                          (totalTime / this.state.questionsAnswered) * 100
+                      ) / 100
                     : 0,
-            finalScore: this.state.score,
         }
     }
 
-    public getState(): GameState {
-        return { ...this.state }
+    protected getGameData(): Record<string, unknown> {
+        return {
+            seenOnePlusOne: this.achievementFlags.seenOnePlusOne,
+            onePlusOneIncorrect: this.achievementFlags.onePlusOneIncorrect,
+            seenOperand999: this.achievementFlags.seenOperand999,
+            zeroAnswerIncorrect: this.achievementFlags.zeroAnswerIncorrect,
+            correctAnswers: this.state.correctAnswers,
+            wrongAnswers: this.state.incorrectAnswers,
+        }
     }
 
-    public isGameActive(): boolean {
-        return this.state.isGameActive
+    // Required abstract methods
+    update(_deltaTime: number): void {
+        // Quick Math doesn't need continuous updates
+        // Timer and scoring are handled by the framework
     }
 
-    public isGameOver(): boolean {
-        return this.state.isGameOver
+    render(): void {
+        // Rendering is handled by the renderer
     }
 
-    public destroy(): void {
-        this.stopGameTimer()
-    }
-
-    // Expose achievement flags for score submission
-    public getAchievementFlags(): {
-        seenOnePlusOne: boolean
-        onePlusOneIncorrect: boolean
-        seenOperand999: boolean
-        zeroAnswerIncorrect: boolean
-    } {
-        return { ...this.achievementFlags }
+    cleanup(): void {
+        // No special cleanup needed for Quick Math
     }
 }
