+++ src/pages/quick-math/index.astro
@@ -216,7 +216,7 @@ const gameNavigation = [
 
     <!-- Quick Math Game Logic -->
     <script>
-      import { initQuickMathFramework } from '@/lib/games/quick-math/FrameworkInit'
+      import { initQuickMathFramework } from '@/lib/games/quick-math/init'
 
       // Initialize game when page loads
       document.addEventListener('DOMContentLoaded', async () => {
